Add White_List Option; Only Display Albums with Required Phrase
This is a bit safer from a privacy perspective as it requires
an affirmative act add a particular phrase to a gallery name
in order for it to be returned here.

The whitelist phrase is removed from the metadata title so it
will not appear in your nanogallery2.

For example, I use the phrase ' - Gallery' as my whitelist,
then any folder I want to display in nanogallery2, I append the
whitelist name to the end of the folder name.